refactor: divided controllers between favorites and notes
