+++ src/routers/favoritesRouter.ts
@@ -1,13 +1,13 @@
 import { Router } from "express";
-import * as notesController from "@/controllers/notesController";
+import * as favoritesController from "@/controllers/favoritesController";
 import { validateParams } from "@/middlewares/validationMiddleware";
 import { noteIdSchema } from "@/middlewares/schemas/notesSchema";
 
 const favoritesRouter = Router();
 
 favoritesRouter
-  .get("/", notesController.getFavoriteNotes)
-  .post("/:id", validateParams(noteIdSchema), notesController.postFavorite)
-  .delete("/:id", validateParams(noteIdSchema), notesController.deleteFavorite)
+  .get("/", favoritesController.getFavoriteNotes)
+  .post("/:id", validateParams(noteIdSchema), favoritesController.postFavorite)
+  .delete("/:id", validateParams(noteIdSchema), favoritesController.deleteFavorite);
 
 export default favoritesRouter;
